+++ example/lib/main.dart
@@ -1,20 +1,10 @@
 import 'package:flutter/material.dart';
-import 'dart:io' if (dart.library.html) 'dart:html' as html;
-import 'dart:typed_data';
-import 'package:flutter/services.dart';
-import 'package:flutter/foundation.dart' show kIsWeb;
-import 'package:image_picker/image_picker.dart';
 import 'package:logging/logging.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:advanced_image_processing_toolkit/advanced_image_processing_toolkit.dart';
-import 'package:advanced_image_processing_toolkit/src/filters.dart';
-import 'package:advanced_image_processing_toolkit/src/object_recognition.dart';
-import 'package:image/image.dart' as img;
-import 'dart:io' show File;
 import 'feature_test.dart';
 
-final _logger = Logger('AdvancedImageProcessingToolkit');
+
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();