+++ test/filters_fallback_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:advanced_image_processing_toolkit/src/filters.dart';
-import 'dart:typed_data';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();